requirements.txt
Django
djangorestframework

dj-database-url

psycopg2-binary

requests

Pillow #para las imagenes